+++ ComicAPI/src/main/java/org/stapledon/caching/ICachable.java
@@ -3,6 +3,9 @@ package org.stapledon.caching;
 
 public interface ICachable
 {
-    // Get the Cache Location
+    /**
+     * Get the path full path where this comic has been cached. Include any augmentation
+     * @return Path
+     */
     String cacheLocation();
 }