+++ ComicAPI/src/main/java/org/stapledon/config/IComicsBootstrap.java
@@ -1,9 +1,13 @@
 package org.stapledon.config;
 
+import org.stapledon.downloader.IDailyComic;
+
 import java.time.LocalDate;
 
 public interface IComicsBootstrap
 {
     String stripName();
     LocalDate startDate();
+
+    IDailyComic getDownloader();
 }